Move resources to main pkg
